homework_03/requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
Faker==9.8.0
pytest==6.2.5
pytest-forked==1.3.0
pytest-xdist==2.4.0
requests==2.25.1
requests-file==1.5.1
requests-ftp==0.3.1
requests-gssapi==1.2.3
selenium==3.141.0
urllib3==1.25.10
webdriver-manager==3.5.1
xdist==0.0.1
